Fix integer/double type mismatch for group index arrays
NIMBLE inferred grp_inds as integer(1) from subscript usage in groupGPvec,
but groupLookup was declared as double(2) everywhere, so slices passed as
grp_inds were double — causing a C++ compile error ('no known conversion
from NimArr<1,double> to NimArr<1,int>').

Fix: declare grp_inds, groupLookup, groupNum and groupNeighbours as integer
in all nimbleFunction signatures and registerDistributions type strings.
Also coerce these arrays to integer in setupGeoMixModel() when building the
NIMBLE data list, since groupLookup is constructed via data.frame pivot and
comes out numeric.

+++ R/nimble_groupVecchiaMarginal.R
@@ -27,7 +27,7 @@
 #' @keywords internal
 groupGPvec_margAlpha <- nimble::nimbleFunction(
   run = function(k = integer(0),
-                 grp_inds = double(1),
+                 grp_inds = integer(1),
                  nb_inds = double(1),
                  X = double(2),
                  m_alpha = double(1),
@@ -188,8 +188,8 @@ evaluateGroupGPvec_margAlpha <- nimble::nimbleFunction(
                  K = integer(0), Z2_ind = double(1),
                  dID = double(1), locID = double(1),
                  distD = double(2), distL = double(2),
-                 m = integer(0), groupLookup = double(2),
-                 groupNum = double(1), groupNeighbours = double(2)) {
+                 m = integer(0), groupLookup = integer(2),
+                 groupNum = integer(1), groupNeighbours = integer(2)) {
     returnType(double(2))
     G <- length(groupNum)
     logdens <- matrix(0, nrow = G, ncol = K)
@@ -275,8 +275,8 @@ dGPgroupvec_margAlpha <- nimble::nimbleFunction(
                  K = integer(0), Z2_ind = double(1),
                  dID = double(1), locID = double(1),
                  distD = double(2), distL = double(2),
-                 m = integer(0), groupLookup = double(2),
-                 groupNum = double(1), groupNeighbours = double(2),
+                 m = integer(0), groupLookup = integer(2),
+                 groupNum = integer(1), groupNeighbours = integer(2),
                  log = integer(0)) {
     returnType(double(0))
     log_terms <- evaluateGroupGPvec_margAlpha(Z2 = x,
@@ -335,8 +335,8 @@ rGPgroupvec_margAlpha <- nimble::nimbleFunction(
                  K = integer(0), Z2_ind = double(1),
                  dID = double(1), locID = double(1),
                  distD = double(2), distL = double(2),
-                 m = integer(0), groupLookup = double(2),
-                 groupNum = double(1), groupNeighbours = double(2)) {
+                 m = integer(0), groupLookup = integer(2),
+                 groupNum = integer(1), groupNeighbours = integer(2)) {
 
     returnType(double(1))
 
@@ -517,8 +517,8 @@ dGPgroupvec_margAlphaP1 <- nimble::nimbleFunction(
                  K = integer(0), Z2_ind = double(1),
                  dID = double(1), locID = double(1),
                  distD = double(2), distL = double(2),
-                 m = integer(0), groupLookup = double(2),
-                 groupNum = double(1), groupNeighbours = double(2),
+                 m = integer(0), groupLookup = integer(2),
+                 groupNum = integer(1), groupNeighbours = integer(2),
                  log = integer(0)) {
     returnType(double(0))
     m_alphaVec <- numeric(1)
@@ -581,8 +581,8 @@ rGPgroupvec_margAlphaP1 <- nimble::nimbleFunction(
                  K = integer(0), Z2_ind = double(1),
                  dID = double(1), locID = double(1),
                  distD = double(2), distL = double(2),
-                 m = integer(0), groupLookup = double(2),
-                 groupNum = double(1), groupNeighbours = double(2)) {
+                 m = integer(0), groupLookup = integer(2),
+                 groupNum = integer(1), groupNeighbours = integer(2)) {
 
     returnType(double(1))
 
@@ -757,7 +757,7 @@ rGPgroupvec_margAlphaP1 <- nimble::nimbleFunction(
 #' @keywords internal
 groupGPvec_margAlphak <- nimble::nimbleFunction(
   run = function(k = integer(0),
-                 grp_inds = double(1),
+                 grp_inds = integer(1),
                  nb_inds = double(1),
                  X = double(2),
                  m_alpha = double(1),
@@ -919,8 +919,8 @@ dGPgroupvec_margAlphak <- nimble::nimbleFunction(
                  Z2_ind = double(1),
                  dID = double(1), locID = double(1),
                  distD = double(2), distL = double(2),
-                 m = integer(0), groupLookup = double(2),
-                 groupNum = double(1), groupNeighbours = double(2),
+                 m = integer(0), groupLookup = integer(2),
+                 groupNum = integer(1), groupNeighbours = integer(2),
                  log = integer(0)) {
     returnType(double(0))
     G <- length(groupNum)
@@ -997,8 +997,8 @@ rGPgroupvec_margAlphak <- nimble::nimbleFunction(
                  Z2_ind = double(1),
                  dID = double(1), locID = double(1),
                  distD = double(2), distL = double(2),
-                 m = integer(0), groupLookup = double(2),
-                 groupNum = double(1), groupNeighbours = double(2)) {
+                 m = integer(0), groupLookup = integer(2),
+                 groupNum = integer(1), groupNeighbours = integer(2)) {
 
     returnType(double(1))
